requirements.txt
streamlit
langchain
langchain-community
langchain-openai
openai
faiss-cpu
pypdf
python-dotenv
tiktoken
chromadb